+++ src/spyglass/utils/dj_mixin.py
@@ -172,7 +172,7 @@ class SpyglassMixin(ExportMixin):
             table_dict[resolved],
         )
 
-    def fetch_nwb(self, *attrs, **kwargs):
+    def fetch_nwb(self, log_export=True, *attrs, **kwargs):
         """Fetch NWBFile object from relevant table.
 
         Implementing class must have a foreign key reference to Nwbfile or
@@ -184,7 +184,7 @@ class SpyglassMixin(ExportMixin):
         """
         table, tbl_attr = self._nwb_table_tuple
 
-        if self.export_id and "analysis" in tbl_attr:
+        if log_export and self.export_id and "analysis" in tbl_attr:
             self._log_fetch_nwb(table, tbl_attr)
 
         return fetch_nwb(self, self._nwb_table_tuple, *attrs, **kwargs)
@@ -462,7 +462,7 @@ class SpyglassMixin(ExportMixin):
             logger.warning(f"Table is empty. No need to delete.\n{self}")
             return
 
-        if self._has_updated_dj_version:
+        if self._has_updated_dj_version and not isinstance(self, dj.Part):
             kwargs["force_masters"] = True
 
         external, IntervalList = self._delete_deps[3], self._delete_deps[4]